requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Bootstrap==3.3.7.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==2.3.7
WTForms==3.0.1
gunicorn==21.2.0
markupsafe==2.1.3